chore(trading): continue with marketsetup
